requirements.txt
django>=3.2, <3.3
djangorestframework>=3.12.4, <3.13.0
django-filter>=2.4.0, <2.5.0
drf-yasg>=1.20.0, <1.21.0
markdown>=3.3.4, <3.4
python-decouple>=3.4, <3.5
psycopg2-binary>=2.8.6, <2.9.0
pyyaml>=5.4.1, <5.5.0
